[cvnirc-qt] Add initial network code
The input from the connect dialog are now used to actually make
a network connection. You'll see potential status messages
on connect, plus an IRC "ERROR" message for timeout potentially
half a minute later.

It seems very strange to me that the line scanning/extraction
needed so much code and logic. There must be an easier way for this.
But it has yet to be found. Perhaps Qt itself could give more/better
abstraction...