[cvnirc-qt] Add LogBuffer: Encapsulate textEditLogbuffer + append method as reuseable widget
It didn't appear in the Designer and I don't know how to add it, there;
so I manually added it to the mainwindow.ui XML. It looks bad at design
time, as the Designer appears to not know at all what size it should
take; but it appears to work fine at run-time. (There's just some
additional margin/padding/spacing, there.)

The idea is to have multiple logbuffers at the same time, later.