[cvnirc-cli, cvnirc-core] Use GNU readline to prompt for connection information, too
This is more complicated code, as it maintains additional state
in the TerminalUI instance, but has the advantage that an existing
connection (/ the whole event loop) may stay running while prompting
for another connection.

Another change is that the connection information is set successively
in the IRCProtoClient instance, so no connectToIRCServer() is used,
but reconnectToIRCServer() directly. This could mess things up
if a reconnect should happen while in the middle of prompting/updating,
but could also be a good thing: The user wanting to update only one part
of the connection information would just press Enter on the parts that
should stay the same, then get reconnected with the updated settings.