[cvnirc-cli] GNU readline: Only queue user input in the callback
See in-file comments in TerminalUI::handle_inNotify_activated().

The essence is that readline seems to react differently when its
functions get called from the callback; so don't do that anymore,
just enqueue lines to later be processed from outside the callback,
with the readline functions hopefully behaving normally again.

This also means that TerminalUI::userInput() can't assume anymore
that it's called from the callback and wouldn't need wrapping of
output operations by calls to readline functions. So its direct use
of "out" had to be replaced by use of outLine().