[cvnirc-gui] MainWindow: Remove icon theming for Connect/Reconnect/Disconnect
For those, the icon theme names were always a poor choice/approximation.
The new iconset should be more self-consistent, so always use that.

An exception is the Quit action, which stays at icon theming, as
the used theme name, application-exit, is semantically correct
and I like the (GTK?) door-with-red-arrow-to-it better than
the quit icon from the new iconset, which is a door before a globe (?),
which is hard to recognize at 32px...