Projects: Add README.txt, TODO.txt as DISTFILES
It did not seem to have any effect in Qt Creator to add those files
to the subdirs project file. So I added them to each subproject,
instead. I hope this will not clash.

The intention was to have these possibly-important documentation files
appear in Qt Creator's Projects window, too; to be easily able to
open them graphically, and to confirm they're part of the project.