Set an overall version, 0.5.3
This seems to be used in, e.g., "make dist".

Now the question arises whether to increase the version number
at begin of development, so that no "stable"/released version
installed on a system will be inadvertantly overwritten by a
development version. Or, another possibility is to increase just
before release/tagging. This avoids the problem of a development
version being installed as (/ "with the looks of") the following
released version.

Perhaps a third option would be best: Having an additional "+dev"
suffix on the previous, or "-dev" suffix on the following, version
number. But I'm not sure if that can be done or how it will be
interpreted numerically, then.