[cvnirc-{cli,core}] Switch CLI to multi-server technology
It can't actually create or handle another connection (yet?),
but it does use the more abstract multi-connection technology, now.
This is able to handle a few context-establishing messages on its own,
which means that cvnirc-cli now supports a few things technically
that it didn't before.

For example, there is technical (though not yet functioning, due to
the lack of switching contexts,) "chat support"; when a line is typed
that does not start with a slash '/', it is treated as a text message
and the CLI attempts to send it as such.

Commands now have to be started with a slash '/', like in the GUI.

* Add IRCCore::createIRCProtoClient(), which adds an IRC protocol client
  without immediately connecting it (as the GUI does), so that the CLI
  can build the connection information on it incrementally.
  .
  In fact, IRCCore::connectToIRCServer() now internally uses it as well,
  or looked at it differently, the create...() code was factored out
  of connectTo...().
* TerminalUI now holds an IRCCore, a current IRCCoreContext and
  a CommandLayer; it just held a single IRCProtoClient, previously.
* The outLine(), outSendingLine() and outReceivedLine() methods/slots
  have been extended to take an optional context argument, formally.
  (Though the central handling method for this does not access the
  context's state information, yet. That still has to be refactored
  from the GUI.) The connectionStateChanged handler now also takes
  an optional context.
* The central "IRCCoreContext got created" handler,
  handle_irc_createdContext(), has been added. With this, the CLI can
  connect all the signals that were previously connected to the single
  IRCProtoClient instance to each new context as it springs into
  existence.
* There is no longer a need to pretend or keep-around-for-further-
  improvement the receivedMessage handler; all that we need should
  already be done by the IRCCoreContext-s.
* The userInputState internal setter and the userInput() method
  needed to be slightly changed to take the IRCProtoClient to modify
  from the current context.

That was it! Sounds really more complicated then it was to
do the changes.