[WIP] [cvnirc-core] CommandGroup: Use std::shared_ptr<CommandDefinition>
It didn't compile before as CommandDefinition is not
default-constructible. This has been circumvented, now, by not storing
instances, but pointers. To avoid getting into trouble with manual
new/delete with something somewhere still holding a reference to it,
use a smartpointer. Apparently this works well with Qt containers
as well.

It still does not compile, though.