Build system: Fix a step that would break in-source directory build
If we're building directly in the source directory, the .qhcp source
file doesn't need to be copied over, first. So don't require the cp
to have been successful.

Similarly, if the source file copy would have been added to QMAKE_CLEAN,
an in-source directory build -> clean would have erroneously removed
the source file.