[cvnirc-gui] MainWindow: Implement starting help viewer (as external process) on our help file
This way, we don't need a build dependency on Qt Help, which on Debian 9
seems to be in package qttools5-dev which does not get pulled in by
package qtbase5-dev, so is less likely to be already installed. (?)
[A skeleton for walking down that path is, as of writing, in branch
wip/QtHelp.]

Besides, it would have been mostly pointless to implement a help viewer
GUI in another window (in the same process) while we mostly don't got
any documentation to show, after all...

The new situation might be sensible, though, for letting casual users
discover about cvnirc-cli and how to switch contexts there. Though this
might not be of any interest to someone clicking the Help menu in a GUI.
...

Well, at least we do have a help infrastructure now, so we could just go
on and fill it with some content. (Patches welcome? Err... Maybe...)