[cvnirc-cli] Parse command line; add options -q, -v, --quiet, --verbose
* Use -q to suppress raw IRC protocol messages, for now.

Use QCommandLineParser, as it seems to have prepared everything for you.

Sadly, it does not seem as if it can be used for /COMMAND arguments
parsing, too, as one main feature, the automatically generated --help
message, seems to be too tailored for the terminal (too spacy for IRC
commands?), and always includes the program name as command name.