[cvnirc-core] IRCProtoClient: Replace buffer handling by a QByteArray-based approach
This operates on a heigher level of abstraction, and seems to work
without need for a round of bugfixes like needed by the approach before.

Sanity checks are done in a different way, now, as the line is first
identified/extracted, only after that it is tested against stray
CRs/LFs, and NULs.

The socket read buffer is now resized dynamically, and may grow
indefinitely, so a check was added to abort the connection when there
is 1 MiB worth of unprocessed line. Actual line length limits
observed on real-world IRC servers should be much smaller.