[WIP] Try to include README.md, TODO.md in the help file as generated HTML
* Remove README.txt & TODO.txt listing as DISTFILES from the
  cvnirc-{core,gui,cli} project files. This was rather a hack
  to get those files listed in Qt Creator's Projects window,
  but we'll get them listed as proper sourcefiles via the doc
  subproject, now.

* doc/doc.pro: Try to set up build from .md to .html for
  README.md & TODO.md; add the resulting README.html & TODO.html
  as additional (but this time not out-of-build-tree) dependencies.
* Add generated README.html & TODO.html to the doc/cvnirc-qt.qhp
  Qt Help Project file.

* Update doc/index.html to link to the generated README.html &
  TODO.html.
* Update doc/build.html to link to the generated README.html,
  and to distinguish between generated README.html and unprocessed
  original README.md.

Sadly, this does not work, yet. The problem seems to be that
qhelpgenerator searches for the files to include in the help file
from where the .qhp input file is located, not from the current working
directory nor from the build directory. (Perhaps making copies again
would be in order. This time perhaps always to a subdirectory,
with name "generated"?)

Another problem: qhelpgenerator seems to ignore any missing files,
they will just give error 404 not found in Qt Assistant, later.
[Update: It couldn't complain about anything missing as it was
configured to include *.html ... The explicit references were
links only, the definition of what files to include is in <files>.]