[WIP] [doc] Build system: Try to influence .depends variables by .variable_out directly
This does not work. The variables stay empty, no dependency is added
to the generated Makefile...

Only the attempt with the INSTALLS worked (somewhat), it just would have
needed an additional .path definition for the file being added. But that
was for the final output file; the stuff not working are the
intermediate file dependencies. (It seems to build fine, but only
through target_predeps. That is, it will be built first if not exist,
but not updated correctly when dependencies changed.)