[doc] Build system: Build Qt help files via OBJECTS instead of via .CONFIG += target_predeps
The latter seemed like a hack to me.

An alternative would have been to add the output to TARGET,
but IIRC the qmake documentation states clearly that multiple TARGETs
are not supported by qmake. Plus, when trying it out all the same
it gives make warnings about overriding/ignoring recipes.