[cvnirc-{gui,cli}] Use PRE_TARGETDEPS to always relink when the internal library changes (on unix)
Theoretically, the generated dependencies on the header files should be
enough to guarantee that the public interface hasn't changed. Plus, it
should generally be possible to replace a shared library on the system
and have the using programs still running. But in practice I sometimes
got misbuilds for which I'm not sure what happened.

Be it like it may, linking again against an updated version of the
internal library should probably do no harm and make sure everything
will still build after cleaning / on an otherwise triggered rebuild.