[cvnirc-core] Exclude normal printable characters from escaping in rawLine signal arguments
This almost restores the previous state, where everything was just
converted to QString; but all non-ASCII or non-printable characters
are now escaped, including the message framing / line terminators.
(The latter is in case we'll support a "compatibility" mode for LF-only
peers (that don't send a proper CR-LF line termiator), one day.
It should be possible then to look at what exactly went over the wire.)

(Perhaps this decision should be revised later, as the %0D%0A at the end
of every protocol dump line can be quite annoying.)

What's important, though, is that:
* protocol dump is now mostly readable again.
* It's possible now to see whether, e.g., a german "Umlaut" had been
  encoded using, e.g., latin1 or UTF-8 on the wire.