[WIP] [cvnirc-core] Introduce IRCProto::make_optional
This should help against the need of spelling out the full
wrapped type in every use of OptionalMessageArgType.

(As this is WIP, included are a previously successful try
of using decltype, which just didn't make the line shorter;
as well as a more general version of the make_optional
template, which didn't work. Both commented out, and to-be-removed
later.)